+++ src/classes/CustomLibMetadata/UI5Parser/UIClass/AbstractUIClass.ts
@@ -3,16 +3,28 @@ export interface UIMethod {
 	params: string[];
 	returnType: string;
 	description: string;
+	api?: string;
 }
 export interface UIField {
 	name: string;
 	type: string | undefined;
-	description: "";
+	description: string;
+}
+export interface UIProperties {
+	name: string;
+	type: string | undefined;
+	description: string;
+}
+export interface UIEvents {
+	name: string;
+	description: string;
 }
 export abstract class AbstractUIClass {
 	public className: string;
 	public methods: UIMethod[] = [];
 	public fields: UIField[] = [];
+	public properties: UIProperties[] = [];
+	public events: UIEvents[] = [];
 	public parentClassNameDotNotation: string = "";
 
 	constructor(className: string, documentText?: string) {