+++ CHANGELOG.md
@@ -35,3 +35,9 @@
 ## 0.8.4 (19-01-2020)
 * Bugfixes
 * Moving params of sap.ui.define is more trustworthy now
+
+## 0.8.5 (31-01-2020)
+* Bugfixes
+* Added parsing of destructured objects in function params
+* Performance improvement for initial completion item generating
+* Dynamic completion items for XML views added. Now autocomplition works for properties and events.